+++ magda-minion-broken-link/src/DevNull.ts
@@ -1,0 +1,11 @@
+import { Writable } from "stream";
+
+export default class DevNull extends Writable {
+    _write(
+        chunk: any,
+        encoding: string,
+        callback: (err?: Error) => void
+    ): void {
+        callback();
+    }
+}